Fix Visual Studio and XCode Warnings

+++ rwengine/src/audio/Sound.hpp
@@ -70,7 +70,7 @@ struct Sound {
         buffer->setMaxDistance(maxDist);
     }
 
-    int getScriptObjectID() const {
+    size_t getScriptObjectID() const {
         return id;
     }
 };
